+++ src/ast/ast.h
@@ -589,6 +589,8 @@ public:
   Expression *expr = nullptr;
   StatementList *stmts = nullptr;
 
+  SizedType ctx_type;
+
 private:
   For(const For &other);
 };